Time: 0 ms (100%), Space: 9.1 MB (57.33%) - LeetHub

+++ 3871-count-commas-in-range-ii/3871-count-commas-in-range-ii.cpp
@@ -3,7 +3,7 @@ public:
     long long countCommas(long long n) {
         long long  result = 0;
         long long lower = 1000;
-        long long comma = 1;
+        int comma = 1;
 
         while(lower <= n){
             long long upper = lower*1000 - 1;
